Bulk fixes for dask implementation of PF
